fix(bloom): ensure wasm32 filter behavior is deterministic
Motivation & Context:
The Bloom filter used `usize` for hash block selection and serialized entry
counts, causing wasm32 to truncate 64-bit hash values and produce different
block locations from x86_64. This could lead to cross-platform false negatives
when filters were created or queried on different targets.

Key Technical Changes:
- Perform block selection with explicit `u64` arithmetic before converting to
  an index, preserving identical hash behavior on 32-bit and 64-bit targets.
- Store block counts as `u32` and entry counts as `u64` for deterministic,
  architecture-independent metadata and serialization.
- Add validation and round-trip tests for Bloom filter bytes, golden hash
  vectors, and wasm32 execution.
- Add the wasm-bindgen test dependency and a CI quality gate covering wasm32
  checks, Clippy, and core unit tests through Node.js.
- Update OCI, worker, and manifest APIs to use the new `u64` entry count type.

Side Effects & Risks:
- Public Bloom filter and manifest APIs changing from `usize` to `u64` require
  callers to update numeric arguments and handle explicit conversions where
  platform-sized values are used.
- The wasm32 CI job adds Rust target, Node.js, and wasm-bindgen CLI setup time.

Verification:
- Added host and wasm-bindgen tests covering hash-location determinism,
  serialization stability, validation errors, and restored-filter lookups.
- CI runs cargo check, Clippy with warnings denied, and wasm32 core tests.

BREAKING CHANGE: Bloom filter and manifest entry-count parameters and accessors
now use `u64` instead of `usize`. Update callers and persisted-data adapters to
use `u64`, adding explicit conversions only at platform-sized allocation
boundaries.

+++ crates/nixcache-core/src/types.rs
@@ -646,7 +646,7 @@ impl ShardDescriptor {
 /// 全局紧凑布隆过滤器元数据容器
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub struct BloomFilterManifest {
-    pub num_entries: usize,
+    pub num_entries: u64,
     pub num_bits: u64,
     pub num_hashes: u8,
     pub blob_digest: String,
@@ -665,7 +665,7 @@ impl BloomFilterManifest {
     }
 
     pub fn new(
-        num_entries: usize,
+        num_entries: u64,
         num_bits: u64,
         num_hashes: u8,
         blob_digest: impl Into<String>,
